python: Split generated code into seperate files
